@@ -1,6 +1,7 @@
+
+import {throwError as observableThrowError, BehaviorSubject, Observable, Subject} from 'rxjs';
 import {Injectable} from '@angular/core';
 import {Headers, Http, RequestOptions, Response} from '@angular/http';
-import {BehaviorSubject, Observable, Subject} from 'rxjs';
 import {ApplicationTaskResult, ApplicationTaskResults} from '../commons/models/application-data-models';
 import {AuthenticationService} from '../commons/services/authentication.service';
 import {SlimLoadingBarService} from 'ng2-slim-loading-bar';
@@ -36,7 +37,7 @@ export class SubscriptionRemoteDataService {
         this.slimLoadingBarService.start();
         this.http.get(this.apiEndpoints['search'], this.getOptions())
             .map((response: Response) => response.json())
-            .catch((error: Response) => Observable.throw({
+            .catch((error: Response) => observableThrowError({
                 success: false,
                 message: 'Error Loading Subscription List',
                 error: error
@@ -84,7 +85,7 @@ export class SubscriptionRemoteDataService {
 
         this.http.get(endPoint, this.getOptions())
             .map((response: Response) => response.json())
-            .catch((error: Response) => Observable.throw({
+            .catch((error: Response) => observableThrowError({
                 success: false,
                 message: 'Error Loading Subscription approval History List',
                 error: error
